+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased](https://github.com/jaivalis/release-raccoon/compare/0.5.0...jdevelop)
 
+- Feat: Added `/me/followed-artists/releases` endpoint
 - Fix: Artist follower count is persisted
 - Fix: Return recommended based on follower count
 - Fix: Align recommended-artists pagination on 0 indexing